Commit tout compile
Il manque les listeners, alertes et deux trois détails dans le main.

+++ src/fr/davidlegras/CustomerState.java
@@ -4,10 +4,10 @@ import java.util.Map;
 
 public interface CustomerState {
 
-    default int price(final Platform platform, final Customer context) {
-        int res = 0;
+    default double price(final Platform platform, final Customer context) {
+        double res = 0;
         for (Map.Entry<Product, Integer> entry : context.cart()) {
-            float productDiscount = 1;
+            double productDiscount = 1;
             for (Offer offer : platform.offers()) {
                 if (offer.applicable(context, entry.getKey()))
                     productDiscount *= (1 - offer.discount());
